ultimate_kg_proj/agentic_graphrag/requirements.txt
# A2A Protocol and Google ADK Dependencies
a2a-sdk[grpc,telemetry]-0.3.0
google-adk>=1.9.0

# Core Dependencies
cognee
python-dotenv>=1.0.0
pydantic>=2.8.0
asyncio>=3.4.3
alembic

# Development Dependencies  
pytest>=8.0.0
pytest-asyncio>=0.24.0
black>=24.0.0
ruff>=0.6.0
mypy>=1.11.0
# Additional Dependencies
uvicorn>=0.30.0
fastapi>=0.115.0
httpx>=0.27.0
aiofiles>=24.1.0